+++ src/main/java/boho/lottonumbergenerator/dto/OfficialLottoSearchRequest.java
@@ -5,8 +5,8 @@ import java.time.LocalDate;
 import org.springframework.format.annotation.DateTimeFormat;
 
 public record OfficialLottoSearchRequest(
-	Long startDrawNumber,
-	Long endDrawNumber,
+	Integer startDrawNumber,
+	Integer endDrawNumber,
 
 	// html lang=ko 일 때 input type date th:field 매핑을 위해서
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
