+++ src/main/java/boho/lottonumbergenerator/dto/OfficialLottoSearchResponse.java
@@ -3,7 +3,7 @@ package boho.lottonumbergenerator.dto;
 import java.time.LocalDate;
 
 public record OfficialLottoSearchResponse(
-	Long drawNumber,
+	Integer drawNumber,
 	LocalDate drawDate,
 	Integer firstPrizeWinnerCount,
 	Long firstPrizeAmount,
